feat: kalshi REST auth_http implementation
