+++ src/adapters/kalshi_rest.rs
@@ -4,31 +4,26 @@
 //! Polymarket-shaped response structs so strategy/execution code can be reused.
 
 use async_trait::async_trait;
-use base64::{engine::general_purpose::STANDARD as BASE64_STANDARD, Engine};
 use chrono::Utc;
-use hmac::{Hmac, Mac};
-use reqwest::header::{HeaderMap, HeaderName, HeaderValue, CONTENT_TYPE};
-use reqwest::{Client, Method};
+use reqwest::Method;
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal::{Decimal, RoundingStrategy};
-use serde_json::{json, Value};
-use sha2::Sha256;
+use serde_json::{Value, json};
 use std::collections::HashMap;
 use tracing::debug;
 
+mod auth_http;
 mod market_data;
 
 use super::{
-    polymarket_clob::OrderBookResponse, BalanceResponse, MarketResponse, MarketSummary,
-    OrderResponse, PositionResponse, TradeResponse,
+    BalanceResponse, MarketResponse, MarketSummary, OrderResponse, PositionResponse, TradeResponse,
+    polymarket_clob::OrderBookResponse,
 };
 use crate::domain::{OrderRequest, OrderSide, OrderStatus};
 use crate::error::{PloyError, Result};
 use crate::exchange::{ExchangeClient, ExchangeKind};
 
 const DEFAULT_KALSHI_API_BASE: &str = "https://api.elections.kalshi.com/trade-api/v2";
-
-type HmacSha256 = Hmac<Sha256>;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum OutcomeSide {
@@ -75,7 +70,7 @@ impl OutcomeSide {
 
 #[derive(Clone)]
 pub struct KalshiClient {
-    http: Client,
+    http: reqwest::Client,
     base_url: String,
     api_key: Option<String>,
     api_secret: Option<String>,
@@ -94,12 +89,7 @@ impl KalshiClient {
             .trim_end_matches('/')
             .to_string();
 
-        let http = Client::builder()
-            .user_agent("ploy-kalshi-adapter/0.1")
-            .build()
-            .map_err(|e| {
-                PloyError::Internal(format!("failed to build Kalshi HTTP client: {}", e))
-            })?;
+        let http = Self::build_http_client()?;
 
         Ok(Self {
             http,
@@ -123,93 +113,6 @@ impl KalshiClient {
 
     pub fn base_url(&self) -> &str {
         &self.base_url
-    }
-
-    fn auth_headers(&self, method: &Method, path: &str, body: &str) -> Result<HeaderMap> {
-        let key = self.api_key.as_ref().ok_or_else(|| {
-            PloyError::Auth("KALSHI_API_KEY (or KALSHI_ACCESS_KEY) is required".to_string())
-        })?;
-        let secret = self.api_secret.as_ref().ok_or_else(|| {
-            PloyError::Auth("KALSHI_API_SECRET (or KALSHI_ACCESS_SECRET) is required".to_string())
-        })?;
-
-        let timestamp = Utc::now().timestamp_millis().to_string();
-        let sign_payload = Self::build_sign_payload(&timestamp, method, path, body);
-        let signature = Self::hmac_signature(secret, &sign_payload)?;
-
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            HeaderName::from_static("kalshi-access-key"),
-            HeaderValue::from_str(key)
-                .map_err(|e| PloyError::Auth(format!("invalid Kalshi API key header: {}", e)))?,
-        );
-        headers.insert(
-            HeaderName::from_static("kalshi-access-signature"),
-            HeaderValue::from_str(&signature)
-                .map_err(|e| PloyError::Auth(format!("invalid Kalshi signature header: {}", e)))?,
-        );
-        headers.insert(
-            HeaderName::from_static("kalshi-access-timestamp"),
-            HeaderValue::from_str(&timestamp)
-                .map_err(|e| PloyError::Auth(format!("invalid Kalshi timestamp header: {}", e)))?,
-        );
-
-        Ok(headers)
-    }
-
-    async fn request_json(
-        &self,
-        method: Method,
-        path: &str,
-        query: Option<&[(&str, String)]>,
-        body: Option<Value>,
-        require_auth: bool,
-    ) -> Result<Value> {
-        let url = format!("{}{}", self.base_url, path);
-        let body_text = body
-            .as_ref()
-            .map(|b| b.to_string())
-            .unwrap_or_else(String::new);
-
-        let mut req = self.http.request(method.clone(), &url);
-
-        if let Some(query) = query {
-            req = req.query(query);
-        }
-
-        if require_auth {
-            let headers = self.auth_headers(&method, path, &body_text)?;
-            req = req.headers(headers);
-        }
-
-        if let Some(body) = body {
-            req = req.header(CONTENT_TYPE, "application/json").json(&body);
-        }
-
-        let resp = req.send().await?;
-        let status = resp.status();
-        let text = resp.text().await?;
-
-        if status.as_u16() == 429 {
-            return Err(PloyError::RateLimited(format!(
-                "Kalshi API rate limited for {} {}",
-                method, path
-            )));
-        }
-
-        if !status.is_success() {
-            return Err(PloyError::Internal(format!(
-                "Kalshi API {} {} failed: status={} body={}",
-                method, path, status, text
-            )));
-        }
-
-        if text.trim().is_empty() {
-            return Ok(Value::Null);
-        }
-
-        serde_json::from_str(&text)
-            .map_err(|e| PloyError::Internal(format!("invalid Kalshi JSON response: {}", e)))
     }
 
     fn pick_array<'a>(root: &'a Value, keys: &[&str]) -> Option<&'a [Value]> {
@@ -243,23 +146,6 @@ impl KalshiClient {
             Value::Number(n) => Decimal::from_str_exact(&n.to_string()).ok(),
             _ => None,
         }
-    }
-
-    fn build_sign_payload(timestamp: &str, method: &Method, path: &str, body: &str) -> String {
-        format!(
-            "{}{}{}{}",
-            timestamp,
-            method.as_str().to_uppercase(),
-            path,
-            body
-        )
-    }
-
-    fn hmac_signature(secret: &str, payload: &str) -> Result<String> {
-        let mut mac = HmacSha256::new_from_slice(secret.as_bytes())
-            .map_err(|e| PloyError::Auth(format!("invalid Kalshi secret: {}", e)))?;
-        mac.update(payload.as_bytes());
-        Ok(BASE64_STANDARD.encode(mac.finalize().into_bytes()))
     }
 
     fn format_price(value: Decimal) -> String {
